requirements.txt
discord.py~=2.4.0
python-dotenv==1.0.0
aiohttp
aiohttp-jinja2
jinja2
prisma
openai>=1.3.7
google-genai
Pillow>=10.4.0
aiofiles
requests
tqdm
colorama
pytz
pgvector